# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mindstrata API",
    "version": "2026-06-05",
    "description": "Проверяемый контракт Go API. Поле x-router-path связывает OpenAPI path с регистрацией mux.HandleFunc в apps/api/internal/httpapi/router.go."
  },
  "servers": [
    {
      "url": "https://mindstrata.ru",
      "description": "Production"
    },
    {
      "url": "https://stage.mindstrata.ru",
      "description": "Staging"
    }
  ],
...
  "paths": {
...
    "/api/profile": {
      "x-router-path": "/api/profile",
      "get": { "operationId": "getProfile", "tags": ["profile"], "security": [{ "cookieAuth": [] }], "responses": { "200": { "$ref": "#/components/responses/AuthUser" }, "401": { "$ref": "#/components/responses/JSONError" } } },
      "delete": { "operationId": "deleteProfile", "tags": ["profile"], "security": [{ "cookieAuth": [] }], "responses": { "200": { "$ref": "#/components/responses/JSONOK" }, "401": { "$ref": "#/components/responses/JSONError" } } }
    },
...
    "/api/admin/promocodes": {
      "x-router-path": "/api/admin/promocodes",
      "get": { "operationId": "getAdminPromocodes", "tags": ["admin"], "security": [{ "cookieAuth": [] }], "responses": { "200": { "$ref": "#/components/responses/List" }, "403": { "$ref": "#/components/responses/JSONError" } } },
      "post": { "operationId": "postAdminPromocodes", "tags": ["admin"], "security": [{ "cookieAuth": [] }], "requestBody": { "$ref": "#/components/requestBodies/PromocodeAdmin" }, "responses": { "201": { "$ref": "#/components/responses/JSONOK" }, "400": { "$ref": "#/components/responses/JSONError" } } }
    },
...
    "/api/admin/payments": {
      "x-router-path": "/api/admin/payments",
      "get": { "operationId": "getAdminPayments", "tags": ["admin", "billing"], "security": [{ "cookieAuth": [] }], "responses": { "200": { "$ref": "#/components/responses/List" }, "403": { "$ref": "#/components/responses/JSONError" } } }
    },
...
    "/webhooks/yookassa/{webhookPath}": {
      "x-router-path": "/webhooks/yookassa/",
      "post": { "operationId": "postYooKassaWebhook", "tags": ["webhook"], "parameters": [{ "name": "webhookPath", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "$ref": "#/components/requestBodies/YooKassaWebhook" }, "responses": { "200": { "$ref": "#/components/responses/JSONOK" }, "400": { "$ref": "#/components/responses/JSONError" }, "403": { "$ref": "#/components/responses/JSONError" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mindstrata_session"
      }
    },
...
    "requestBodies": {
...
      "PromocodeAdmin": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PromocodeAdminRequest" } } } },
...
      "YooKassaWebhook": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/YooKassaWebhook" } } } }
    },
    "responses": {
      "JSONOK": {
        "description": "JSON response with ok=true on success",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OKResponse" } } }
      },
      "JSONError": {
        "description": "JSON error response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "AuthUser": {
        "description": "Authenticated user response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthUserResponse" } } }
      },
...
      "List": {
        "description": "List response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListResponse" } } }
      }
    },
    "schemas": {
      "OKResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": { "ok": { "type": "boolean", "const": true } },
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "type": "boolean", "const": false },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "errorCode": { "type": "string" },
          "requiresAuth": { "type": "boolean" }
        },
        "additionalProperties": true
      },
      "User": {
        "type": "object",
        "required": ["id", "role"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "email": { "type": ["string", "null"], "format": "email" },
          "role": { "type": "string", "enum": ["owner", "admin", "billing_admin", "content_admin", "support", "tester", "user"] },
          "status": { "type": "string" }
        },
        "additionalProperties": true
      },
      "AuthUserResponse": {
        "type": "object",
        "required": ["ok", "user"],
        "properties": {
          "ok": { "type": "boolean" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "additionalProperties": true
      },
...
      "ListResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "ok": { "type": "boolean" },
          "items": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        },
        "additionalProperties": true
      },
...
      "PromocodeAdminRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": { "type": "string", "minLength": 1 },
          "grantsType": { "type": "string" },
          "targetId": { "type": "integer", "format": "int64" },
          "maxUses": { "type": "integer", "minimum": 1 },
          "active": { "type": "boolean" }
        },
        "additionalProperties": true
      },
...
      "YooKassaWebhook": {
        "type": "object",
        "required": ["event", "object"],
        "properties": {
          "event": { "type": "string" },
          "object": { "type": "object", "required": ["id"], "properties": { "id": { "type": "string" }, "status": { "type": "string" } }, "additionalProperties": true }
        },
        "additionalProperties": true
      }
    }
  }
}